+++ src/main/java/step2/game/LottoGame.java
@@ -1,4 +1,4 @@
-package step2.Model;
+package step2.game;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -8,9 +8,9 @@ import java.util.stream.IntStream;
 import static java.util.stream.Collectors.toList;
 
 public class LottoGame {
-    public static final int NUMBER_COUNT = 6;
-    public static final int MIN_NUMBER = 1;
-    public static final int MAX_NUMBER = 45;
+    static final int NUMBER_COUNT = 6;
+    static final int MIN_NUMBER = 1;
+    static final int MAX_NUMBER = 45;
 
     private static final List<Integer> balls = IntStream.rangeClosed(MIN_NUMBER, MAX_NUMBER)
             .boxed().collect(toList());
