+++ src/main/java/step2/game/WinningLotto.java
@@ -1,4 +1,4 @@
-package step2.Model;
+package step2.game;
 
 import java.util.Arrays;
 import java.util.List;
