+++ src/test/java/step2/analyze/PrizeTest.java
@@ -1,9 +1,8 @@
-package step2;
+package step2.analyze;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import step2.Data.Prize;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
