+++ src/test/java/step2/game/LottoGameTest.java
@@ -1,9 +1,8 @@
-package step2;
+package step2.game;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step2.Model.LottoGame;
 
 import java.util.Arrays;
 import java.util.List;
