+++ src/test/java/step2/game/TicketTest.java
@@ -1,8 +1,7 @@
-package step2;
+package step2.game;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step2.Model.Ticket;
 
 import static org.assertj.core.api.Assertions.assertThat;
